Move `type_str` and `replace_str` to `type` and `replace` commands
