+++ yazi-core/src/input/commands/type_.rs
@@ -1,4 +1,5 @@
 use yazi_config::keymap::Key;
+use yazi_macro::render;
 use yazi_shared::event::CmdCow;
 
 use crate::input::{Input, InputMode};
@@ -33,5 +34,18 @@ impl Input {
 		}
 
 		false
+	}
+
+	pub fn type_str(&mut self, s: &str) {
+		let snap = self.snaps.current_mut();
+		if snap.cursor < 1 {
+			snap.value.insert_str(0, s);
+		} else {
+			snap.value.insert_str(snap.idx(snap.cursor).unwrap(), s);
+		}
+
+		self.move_(s.chars().count() as isize);
+		self.flush_value();
+		render!();
 	}
 }
