+++ yazi-fm/src/app/app.rs
@@ -92,8 +92,7 @@ impl App {
 			let input = &mut self.cx.input;
 			if input.mode() == InputMode::Insert {
 				input.type_str(&str);
-			}
-			if input.mode() == InputMode::Replace {
+			} else if input.mode() == InputMode::Replace {
 				input.replace_str(&str);
 			}
 		}
